requirements.txt
flask
flask-cors
pandas
Pillow
tqdm
numpy
torch
torchvision
timm
faiss-cpu